+++ src/components/tab.jsx
@@ -1,11 +1,17 @@
-import { Fragment, useState, useRef } from "react";
-import { Excalidraw, MainMenu, serializeAsJSON } from "@zuoez02/excalidraw";
-import { saveData } from "../services/data";
+import { Fragment, useState, useRef, useEffect } from "react";
+import {
+  Excalidraw,
+  MainMenu,
+  serializeAsJSON,
+  serializeLibraryAsJSON,
+  useHandleLibrary,
+} from "@zuoez02/excalidraw";
+import { loadLibraryFile, saveData, saveLibraryFile } from "../services/data";
 import PropTypes from "prop-types";
 
 export const Tab = (props) => {
   const initData = props.initData;
-  const [realName, setRealName] = useState(props.name);
+  const realName = props.name;
   const [isFullScreen, setIsFullScreen] = useState(false);
   const excalidrawRef = useRef(null);
 
@@ -42,11 +48,21 @@ export const Tab = (props) => {
     saveData(realName, json);
   }
 
+  const handleLibraryChange = (items) => {
+    saveLibrary(items);
+  };
+
+  function saveLibrary(items) {
+    const json = serializeLibraryAsJSON(items);
+    saveLibraryFile(json);
+  }
+
   function save() {
     debounce(() => {
       saveNow();
     }, 500);
   }
+
   return (
     <Fragment>
       <div className="excalidraw-component">
@@ -54,7 +70,8 @@ export const Tab = (props) => {
           ref={excalidrawRef}
           onChange={() => save()}
           langCode="zh-CN"
-          initialData={initData}
+          initialData={{ ...initData }}
+          onLibraryChange={handleLibraryChange}
           onLinkOpen={(element, event) => {
             const link = element.link;
             const { nativeEvent } = event.detail;
@@ -83,22 +100,12 @@ export const Tab = (props) => {
             <MainMenu.DefaultItems.Help />
             <MainMenu.Item
               key="refresh"
-              icon={
-                <svg>
-                  <use xlinkHref="#iconHelp"></use>
-                </svg>
-              }
               onSelect={() => window.location.reload(true)}
             >
               刷新
             </MainMenu.Item>
             <MainMenu.Item
               key="fullscreen"
-              icon={
-                <svg>
-                  <use xlinkHref="#iconHelp"></use>
-                </svg>
-              }
               onSelect={() => toggleFullscreen()}
             >
               全屏