+++ src/utils/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { loadData, loadAllFiles } from "../services/data";
+import { loadData, loadAllFiles, loadLibraryFile } from "../services/data";
 import { Dock } from "../components/dock";
 import { Tab } from "../components/tab";
 
@@ -24,10 +24,11 @@ export function initExcalidrawTab(plugin) {
       this.element.innerHTML =
         '<div class="fn__flex fn__flex-1 fn__flex-column"><div style="border: none" class="excalidraw excalidraw-wrapper fn__flex fn__flex-1"></div></div>';
       const initData = await loadData(this.data.name);
+      const libraryItems = (await loadLibraryFile())?.libraryItems || [];
       const root = ReactDOM.createRoot(
         this.element.querySelector(".excalidraw-wrapper")
       );
-      root.render(React.createElement(Tab, { initData, name: this.data.name, el: this.element }));
+      root.render(React.createElement(Tab, { initData: { ...initData, libraryItems }, name: this.data.name, el: this.element }));
       this.data.destroy = () => {
         root && root.unmount();
       };